reset clouds (and get rid of cloudparams.h)
