+++ src/skyparams.h
@@ -68,6 +68,16 @@ struct StarParams
 	f32 scale;
 };
 
+struct CloudParams
+{
+	float density;
+	video::SColor color_bright;
+	video::SColor color_ambient;
+	float thickness;
+	float height;
+	v2f speed;
+};
+
 // Utility class for setting default sky, sun, moon, stars values:
 class SkyboxDefaults
 {
@@ -130,5 +140,17 @@ public:
 		stars.starcolor = video::SColor(105, 235, 235, 255);
 		stars.scale = 1;
 		return stars;
+	}
+
+	const CloudParams getCloudDefaults()
+	{
+		CloudParams clouds;
+		clouds.density = 0.4f;
+		clouds.color_bright = video::SColor(229, 240, 240, 255);
+		clouds.color_ambient = video::SColor(255, 0, 0, 0);
+		clouds.thickness = 16.0f;
+		clouds.height = 120;
+		clouds.speed = v2f(0.0f, -2.0f);
+		return clouds;
 	}
 };
